# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для тестирования стажёров Активбанка через Google Forms.",
        "title": "Activ GForms API",
        "contact": {
            "name": "Amin Muborakkadamov",
            "url": "https://github.com/AminVamar"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/branches": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "филиалы"
                ],
                "summary": "Филиалы",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.BranchesResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "филиалы"
                ],
                "summary": "Добавить филиал",
                "parameters": [
                    {
                        "description": "филиал",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateBranchRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Branch"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/tests": {
            "get": {
                "description": "Список тестов из внешнего источника.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "тесты"
                ],
                "summary": "Все тесты из источника",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.TestsResponse"
                        }
                    },
                    "503": {
                        "description": "источник тестов недоступен",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/tests/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "тесты"
                ],
                "summary": "Тест по идентификатору",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Test"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "источник тестов недоступен",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Branch": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "domain.Option": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_correct": {
                    "type": "boolean"
                },
                "question_id": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "domain.Question": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Option"
                    }
                },
                "type": {
                    "type": "integer"
                }
            }
        },
...
        "domain.Test": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Question"
                    }
                },
                "test_type": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "http.BranchesResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Branch"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "http.CreateBranchRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "6900"
                },
                "name": {
                    "type": "string",
                    "example": "ЦБО Рудаки"
                }
            }
        },
...
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "http.TestsResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Test"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}